refactor: extract SequencerGridHelpers (603→474 lines)
New file: src/components/sequencer/SequencerGridHelpers.tsx (155 lines)
- VelocityLane, SamplePickerDropdown

SequencerGrid.tsx: 603 → 474 lines (21% reduction)
All 5 oversized components now under 600 lines ✅
Build: 0 errors. Tests: 46/46 pass.
